Introduce EditorConfig
Most editors respect EditorConfig out-of-the-box, so it's useful to tell
editors the indentation.  However, it cannot tell editors more complex
style configuration such as quote type, trailing comma, semicolon, etc.,
which are language-specific and/or need syntax analysis, and currently
are done by formatters.  (There are proposals to extend EditorConfig to
configure formatter, linter, and language server options, but they're
not implemented yet.)